backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pdfplumber==0.10.3
PyMuPDF==1.23.8
python-docx==1.1.0
reportlab==4.0.7
anthropic==0.39.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.1
python-dotenv==1.0.0
pydantic==2.5.2
pydantic-settings==2.1.0
